refactor: pagination component
